requirements.txt
Django==5.0.7
djangorestframework==3.15.2
google-cloud-logging==3.10.0
python-dotenv==1.0.1